Self PR review - moves format logic to shared folder so front and back end can use same logic/package
