+++ packages/back-end/package.json
@@ -100,7 +100,6 @@
     "saslprep": "^1.0.0",
     "shared": "0.0.1",
     "snowflake-sdk": "1.10.1",
-    "sql-formatter": "^11.0.2",
     "ssrf-req-filter": "^1.1.1",
     "string-env-interpolation": "^1.0.1",
     "stripe": "^13.1.0",
