+++ packages/back-end/src/util/sql.ts
@@ -1,5 +1,5 @@
-import { format as sqlFormat, FormatOptions } from "sql-formatter";
 import Handlebars from "handlebars";
+import { FormatDialect, format } from "shared";
 import { SQLVars } from "back-end/types/sql";
 import {
   FactTableColumnType,
@@ -11,6 +11,9 @@ import { helpers } from "./handlebarsHelpers";
 Object.keys(helpers).forEach((helperName) => {
   Handlebars.registerHelper(helperName, helpers[helperName]);
 });
+
+// Re-export FormatDialect and format for backward compatibility
+export { FormatDialect, format };
 
 export function getBaseIdTypeAndJoins(
   objects: string[][],
@@ -162,19 +165,6 @@ export function compileSqlTemplate(
       );
     }
     throw new Error(`Error compiling SQL template: ${e.message}`);
-  }
-}
-
-export type FormatDialect = FormatOptions["language"] | "";
-export function format(sql: string, dialect?: FormatDialect) {
-  if (!dialect) return sql;
-
-  try {
-    return sqlFormat(sql, {
-      language: dialect,
-    });
-  } catch (e) {
-    return sql;
   }
 }
