+++ packages/front-end/package.json
@@ -87,7 +87,6 @@
     "scim2-parse-filter": "^0.2.8",
     "shared": "0.0.1",
     "spectacle": "^9.7.2",
-    "sql-formatter": "^15.6.2",
     "swr": "^2.2.5",
     "uniqid": "^5.3.0",
     "uuid": "^9.0.0",
