+++ packages/shared/package.json
@@ -23,6 +23,7 @@
     "date-fns": "^2.15.0",
     "dirty-json": "^0.9.2",
     "json-stringify-pretty-compact": "^3.0.0",
+    "sql-formatter": "^11.0.2",
     "uniqid": "^5.4.0",
     "zod": "^3.22.4"
   },
