+++ packages/shared/src/index.ts
@@ -8,3 +8,4 @@ export * as sdkVersioning from "./sdk-versioning";
 export * as types from "./types";
 export * as power from "./power";
 export * as health from "./health";
+export * from "./types/sql";
